airframe-http: Embed airframe-http-router project into airframe-http (JVM)
