+++ airframe-http-codegen/src/main/scala/wvlet/airframe/http/codegen/HttpClientIR.scala
@@ -19,10 +19,10 @@ import wvlet.airframe.http.{HttpMethod, Router}
 import wvlet.airframe.http.codegen.RouteAnalyzer.RouteAnalysisResult
 import wvlet.airframe.http.codegen.client.HttpClientGenerator
 import wvlet.airframe.http.codegen.client.HttpClientGenerator.fullTypeNameOf
-import wvlet.airframe.http.router.Route
 import wvlet.airframe.rx.{Rx, RxStream}
 import wvlet.airframe.surface.{GenericSurface, HigherKindedTypeSurface, MethodParameter, Parameter, Surface, TypeName}
 import wvlet.log.LogSupport
+import wvlet.airframe.http.router.{Route, HttpRequestMapper}
 
 /**
   * Generate an intermediate representation (IR) of Scala HTTP client code from a given airframe-http interface
