+++ airframe-http-finagle/src/main/scala/wvlet/airframe/http/finagle/FinagleServer.scala
@@ -25,8 +25,8 @@ import wvlet.airframe.codec.MessageCodec
 import wvlet.airframe.control.MultipleExceptions
 import wvlet.airframe.http.finagle.FinagleServer.FinagleService
 import wvlet.airframe.http.finagle.filter.HttpAccessLogFilter
-import wvlet.airframe.http.router.{ControllerProvider, ResponseHandler}
 import wvlet.airframe.http.{HttpBackend, HttpHeader, HttpMessage, HttpServerException, RPCContext, RPCException, Router}
+import wvlet.airframe.http.router.{ControllerProvider, HttpRequestDispatcher, ResponseHandler}
 import wvlet.airframe.surface.Surface
 import wvlet.log.LogSupport
 import wvlet.log.io.IOUtil
