+++ airframe-http-grpc/src/main/scala/wvlet/airframe/http/grpc/internal/GrpcServiceBuilder.scala
@@ -26,7 +26,7 @@ import wvlet.airframe.http.grpc.{
   GrpcServerConfig,
   GrpcService
 }
-import wvlet.airframe.http.router.Route
+import wvlet.airframe.http.router.{Route, HttpRequestMapper}
 import wvlet.airframe.msgpack.spi.MsgPack
 import wvlet.airframe.surface.{MethodParameter, MethodSurface, Surface, TypeName}
 import wvlet.airframe.rx._
