+++ airframe-http-grpc/src/test/scala/wvlet/airframe/http/grpc/example/Greeter.scala
@@ -16,8 +16,8 @@ import io.grpc.stub.{AbstractBlockingStub, ClientCalls}
 import io.grpc.{CallOptions, Channel}
 import wvlet.airframe.codec.MessageCodecFactory
 import wvlet.airframe.http.grpc.internal.GrpcServiceBuilder
-import wvlet.airframe.http.router.Route
 import wvlet.airframe.http.{RPC, Router}
+import wvlet.airframe.http.router.Route
 
 /**
   */
