+++ airframe-http-netty/src/main/scala/wvlet/airframe/http/netty/NettyResponseHandler.scala
@@ -15,8 +15,8 @@ package wvlet.airframe.http.netty
 
 import wvlet.airframe.codec.{JSONCodec, MessageCodec, MessageCodecFactory}
 import wvlet.airframe.http.HttpMessage.{Request, Response}
-import wvlet.airframe.http.router.{ResponseHandler, Route}
 import wvlet.airframe.http.{Http, HttpStatus}
+import wvlet.airframe.http.router.{ResponseHandler, Route}
 import wvlet.airframe.msgpack.spi.MsgPack
 import wvlet.airframe.surface.{Primitive, Surface}
 import wvlet.log.LogSupport
