--- airframe-http-router/src/test/scala/wvlet/airframe/http/InheritedEndpointTest.scala
@@ -1,17 +1,0 @@
-package wvlet.airframe.http
-
-import wvlet.airspec._
-
-object InheritedEndpointTest extends AirSpec {
-
-  trait Base {
-    @Endpoint(path = "/hello")
-    def hello: String = "hello"
-  }
-  trait MyApp extends Base
-
-  test("find inherited endpoints") {
-    val router = Router.of[MyApp]
-    router.routes.find(_.path == "/hello") shouldBe defined
-  }
-}
