+++ airframe-http/.jvm/src/main/scala/wvlet/airframe/http/HttpAccessLogWriter.scala
@@ -15,9 +15,9 @@ package wvlet.airframe.http
 
 import wvlet.airframe.codec.MessageCodec
 import wvlet.airframe.http.internal.RPCCallContext
-import wvlet.airframe.surface.{MethodSurface, Parameter, Surface, TypeName}
+import wvlet.airframe.surface.{Parameter, Surface, TypeName}
 import wvlet.airframe.ulid.ULID
-import wvlet.log.{AsyncHandler, LogFormatter, LogRecord, LogRotationHandler, LogTimestampFormatter}
+import wvlet.log._
 
 import java.lang.reflect.InvocationTargetException
 import java.util.Locale
