+++ airframe-http/.jvm/src/main/scala/wvlet/airframe/http/Router.scala
@@ -13,10 +13,11 @@
  */
 package wvlet.airframe.http
 
-import wvlet.airframe.http.router._
-import wvlet.airframe.surface.{GenericSurface, HigherKindedTypeSurface, MethodSurface, Surface, TypeName}
+import wvlet.airframe.http.router.Automaton.DFA
+import wvlet.airframe.surface._
 import wvlet.log.LogSupport
 
+import wvlet.airframe.http.router.{Route, RouteMatcher, ControllerRoute, RouteMatch}
 import scala.annotation.tailrec
 import scala.language.experimental.macros
 import scala.language.higherKinds
@@ -228,7 +229,7 @@ case class Router(
   }
 }
 
-object Router extends RouterObjectBase with LogSupport {
+object Router extends router.RouterObjectBase with LogSupport {
   val empty: Router   = new Router()
   def apply(): Router = empty
